backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Agent Framework
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# LLM
google-generativeai>=0.8.0

# Data & Tools
yfinance>=0.2.35
pandas>=2.0.0
numpy>=1.24.0

# Search & Scraping
ddgs>=0.1.0

# PDF Generation
reportlab>=4.0.0

# Auth & Database
pymongo>=4.6.0
PyJWT>=2.8.0
bcrypt>=4.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.27.0

# Testing
pytest>=7.0.0